dental ai aget/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1

langchain==0.1.16
langchain-core==0.1.48
langchain-community==0.0.34
langchain-google-genai==0.0.11

google-generativeai==0.5.2

mysql-connector-python==8.3.0
requests==2.31.0
pydantic==2.6.4
twilio==8.1.0

